feat(#2415): split out appearance diagnostics
